fixing bug of duplication when replying

+++ hx_lti_initializer/static/Hxighlighter/hxighlighter_text.js
@@ -1,4 +1,4 @@
-// [AIV_SHORT]  Version: 0.0.1 - Monday, September 9th, 2019, 2:30:44 PM  
+// [AIV_SHORT]  Version: 0.0.1 - Monday, September 9th, 2019, 2:34:26 PM  
  /******/ (function(modules) { // webpackBootstrap
 /******/ 	// The module cache
 /******/ 	var installedModules = {};
@@ -37565,41 +37565,45 @@ __webpack_require__(65);
     var reply_list = jQuery(viewer).find('.plugin-area-bottom div[class*=reply-list]');
 
     if (reply_list.is(':visible')) {
-      reply_list.append("<div class='reply reply-item-" + reply.id + "'>" + delete_option + "<strong>" + reply.creator.name + "</strong> (" + jQuery.timeago(reply.created) + "):" + reply.annotationText.join('<br><br>') + "</div>");
-      jQuery('.reply.reply-item-' + reply.id + ' .delete-reply').confirm({
-        'title': 'Delete Reply?',
-        'content': 'Would you like to delete your reply? This is permanent.',
-        'buttons': {
-          confirm: function confirm() {
-            $.publishEvent('StorageAnnotationDelete', self.instanceID, [reply]);
-            annotation.replies = annotation.replies.filter(function (ann) {
-              if (ann.id !== reply.id) {
-                return ann;
+      if (reply_list.find('reply-item-' + reply.id).length > 0) {
+        reply_list.find('reply-item-' + reply.id + ' .reply-body').html(reply.annotationText.join('<br>'));
+      } else {
+        reply_list.append("<div class='reply reply-item-" + reply.id + "'>" + delete_option + "<strong>" + reply.creator.name + "</strong> (" + jQuery.timeago(reply.created) + "):" + reply.annotationText.join('<br><br>') + "</div>");
+        jQuery('.reply.reply-item-' + reply.id + ' .delete-reply').confirm({
+          'title': 'Delete Reply?',
+          'content': 'Would you like to delete your reply? This is permanent.',
+          'buttons': {
+            confirm: function confirm() {
+              $.publishEvent('StorageAnnotationDelete', self.instanceID, [reply]);
+              annotation.replies = annotation.replies.filter(function (ann) {
+                if (ann.id !== reply.id) {
+                  return ann;
+                }
+              });
+              annotation.totalReplies--;
+
+              annotation._local.highlights.forEach(function (high) {
+                jQuery(high).data('annotation', annotation);
+              });
+
+              jQuery('.reply.reply-item-' + reply.id).remove();
+              jQuery('.side.ann-item.item-' + annotation.id + ' .view-replies').html('View ' + self.pluralize(annotation.totalReplies, 'Reply', 'Replies'));
+
+              if (annotation.totalReplies == 0) {
+                jQuery('.side.ann-item.item-' + annotation.id + ' .create-reply').show();
+                jQuery('.side.ann-item.item-' + annotation.id + ' .view-replies').hide();
+              } else {
+                jQuery('.side.ann-item.item-' + annotation.id + ' .create-reply').hide();
+                jQuery('.side.ann-item.item-' + annotation.id + ' .view-replies').show();
               }
-            });
-            annotation.totalReplies--;
 
-            annotation._local.highlights.forEach(function (high) {
-              jQuery(high).data('annotation', annotation);
-            });
-
-            jQuery('.reply.reply-item-' + reply.id).remove();
-            jQuery('.side.ann-item.item-' + annotation.id + ' .view-replies').html('View ' + self.pluralize(annotation.totalReplies, 'Reply', 'Replies'));
-
-            if (annotation.totalReplies == 0) {
-              jQuery('.side.ann-item.item-' + annotation.id + ' .create-reply').show();
-              jQuery('.side.ann-item.item-' + annotation.id + ' .view-replies').hide();
-            } else {
-              jQuery('.side.ann-item.item-' + annotation.id + ' .create-reply').hide();
-              jQuery('.side.ann-item.item-' + annotation.id + ' .view-replies').show();
-            }
-
-            jQuery('.side.ann-item.item-' + annotation.id).find('.plugin-area-bottom div[class*=reply-list]').hide();
-            jQuery('.side.ann-item.item-' + annotation.id).find('.plugin-area-bottom .reply-menu').hide();
-          },
-          cancel: function cancel() {}
-        }
-      });
+              jQuery('.side.ann-item.item-' + annotation.id).find('.plugin-area-bottom div[class*=reply-list]').hide();
+              jQuery('.side.ann-item.item-' + annotation.id).find('.plugin-area-bottom .reply-menu').hide();
+            },
+            cancel: function cancel() {}
+          }
+        });
+      }
     } else {
       console.log("not visible");
       jQuery('.reply-area-' + annotation.id + ' .view-replies').html('View ' + annotation.totalReplies + ' Replies');
